refactor: svgr export files from src/components/icons to src/icons
