--- apps/client/src/components/icons/ChevronDown.ts
@@ -1,1 +1,0 @@
-export { default as ChevronDownIcon } from '@/assets/common/chevron-down.svg';
