+++ apps/client/src/icons/Chevron.ts
@@ -1,0 +1,1 @@
+export { default as ChevronDownIcon } from '@/assets/common/chevron-down.svg';
