--- apps/client/src/icons/ObjectBoard.ts
@@ -1,2 +1,0 @@
-export { default as ChevronIcon } from '@/assets/object-board/chevron.svg';
-export { default as FilterIcon } from '@/assets/object-board/filter.svg';
